+++ client/components/ModalWithForm.jsx
@@ -1,9 +1,7 @@
 import React from 'react'
 import { Modal, Button } from 'react-bootstrap'
-import { connect } from 'react-redux'
-import * as actions from '../actions'
 
-class ModalWithFormComponent extends React.Component {
+export class ModalWithForm extends React.Component {
     constructor(props) {
         super(props)
     }
@@ -18,9 +16,9 @@ class ModalWithFormComponent extends React.Component {
             ref: 'form'
         })
         return (
-            <Modal show={this.props.show} onHide={this.props.handleHide}>
+            <Modal show={this.props.show} onHide={this.props.onHide}>
                 <Modal.Header>
-                    <a className="close" onClick={this.props.handleHide}>
+                    <a className="close" onClick={this.props.onHide}>
                         <i className="icon-close-small" />
                     </a>
                     <h3>{ this.props.title }</h3>
@@ -29,7 +27,7 @@ class ModalWithFormComponent extends React.Component {
                     { form }
                 </Modal.Body>
                 <Modal.Footer>
-                    <Button onClick={this.props.handleHide}>Close</Button>
+                    <Button onClick={this.props.onHide}>Close</Button>
                     <Button type="submit"
                             onClick={this.submit.bind(this)}
                             disabled={this.props.pristine ||
@@ -40,18 +38,12 @@ class ModalWithFormComponent extends React.Component {
     }
 }
 
-ModalWithFormComponent.propTypes = {
+ModalWithForm.propTypes = {
     form: React.PropTypes.func,
     initialValues: React.PropTypes.object,
     title: React.PropTypes.string,
     show: React.PropTypes.bool,
-    handleHide: React.PropTypes.func,
+    onHide: React.PropTypes.func,
     pristine: React.PropTypes.bool,
     submitting: React.PropTypes.func,
-
 }
-const mapDispatchToProps = (dispatch) => ({
-    handleHide: () => dispatch(actions.hideModal())
-})
-
-export const ModalWithForm = connect(null, mapDispatchToProps)(ModalWithFormComponent)
