+++ client/components/index.js
@@ -1,4 +1,5 @@
 export { EventsList } from './EventsList'
 export { PlanningItem } from './PlanningItem'
+export { ModalWithForm } from './ModalWithForm'
 import * as fields from './fields'
 export { fields }
