+++ client/containers/CreateAgendaContainer.jsx
@@ -1,23 +1,31 @@
 import React from 'react'
-import { ModalWithForm, CreateAgendaForm } from './index'
+import {CreateAgendaForm } from './index'
+import { ModalWithForm } from '../components'
 import { connect } from 'react-redux'
+import * as actions from '../actions'
 
 /**
 * Modal for adding an agenda
 */
-export const CreateAgenda = ({ show }) => (
+export const CreateAgenda = ({ show, handleHide }) => (
     <ModalWithForm
         title="Create an Agenda"
+        onHide={handleHide}
         form={CreateAgendaForm}
         show={show} />
 )
 
 CreateAgenda.propTypes = {
-    show: React.PropTypes.bool
+    show: React.PropTypes.bool,
+    handleHide: React.PropTypes.func,
 }
 
 const mapStateToProps = (state) => ({
     show: state.modal.modalType === 'CREATE_AGENDA',
 })
 
-export const CreateAgendaContainer = connect(mapStateToProps, null)(CreateAgenda)
+const mapDispatchToProps = (dispatch) => ({
+    handleHide: () => dispatch(actions.hideModal())
+})
+
+export const CreateAgendaContainer = connect(mapStateToProps, mapDispatchToProps)(CreateAgenda)
