+++ client/containers/index.js
@@ -8,7 +8,6 @@ export { RepeatEventForm } from './RepeatEventForm'
 export { AddEventContainer } from './AddEventContainer'
 export { CreateAgendaContainer } from './CreateAgendaContainer'
 export { CreateAgendaForm } from './CreateAgendaForm'
-export { ModalWithForm } from './ModalWithForm'
 export { SelectAgenda } from './SelectAgenda'
 export { EditPlanningPanelContainer } from './EditPlanningPanelContainer'
 export { PlanningForm } from './PlanningForm'
